cache siege stats

+++ cogs/utility.py
@@ -6,16 +6,19 @@ from modules import checks, utils, siege
 import json
 import asyncio, aiohttp
 import requests
+import time
 from lxml import html
 from urllib.parse import quote
 from http.client import responses
 from googletrans import Translator, LANGUAGES, LANGCODES
 
-
 class Utility:
     def __init__(self, bot):
         self.bot = bot
         self.translator = Translator()
+
+        self.SIEGE_CACHE = {}
+        self.bot.loop.create_task(self.remove_siege_cache())
         
     @commands.command(description="info about a Discord user", brief="info about a Discord user", pass_context=True)
     @commands.cooldown(2, 5, commands.BucketType.user)
@@ -352,7 +355,7 @@ Created at {date}
                       pass_context=True,
                       aliases=["r6s", "r6stats"])
     @commands.cooldown(1, 5, commands.BucketType.server)
-    async def siege(self, ctx, username : str, stats_selection : str="all", platform : str="uplay"):
+    async def siege(self, ctx, username : str, stats_selection : str="overall", platform : str="uplay"):
         stats_options = ["overall", "ranked", "casual", "all"]
         if (stats_selection not in stats_options):
             await self.bot.messaging.reply(ctx.message, "Invalid stat selection `{}`, options are: {}".format(stats_selection,
@@ -364,37 +367,41 @@ Created at {date}
                 ", ".join("`{}`".format(s) for s in siege.platforms.keys())))
             return
 
-        msg = await self.bot.messaging.reply(ctx.message, "Searching for stats (might take a while)...")
-        await self.bot.send_typing(ctx.message.channel)
+        if (username in self.SIEGE_CACHE):
+            stats = self.SIEGE_CACHE[username]["stats"]
+            self.SIEGE_CACHE[username]["time"] = time.time()
+        else:
+            msg = await self.bot.messaging.reply(ctx.message, "Searching for stats (might take a while)...")
+            await self.bot.send_typing(ctx.message.channel)
 
-        stats = await siege.get_player(username, platform=platform)
+            stats = await siege.get_player(username, platform=platform)
 
-        await self.bot.bot_utils.delete_message(msg)
+            await self.bot.bot_utils.delete_message(msg)
 
-        if (not stats):
-            await self.bot.messaging.reply(ctx.message, "Failed to find `{}` stats for `{}` on `{}`".format(stats_selection, username, platform))
-            return
+            if (not stats):
+                await self.bot.messaging.reply(ctx.message, "Failed to find `{}` stats for `{}` on `{}`".format(stats_selection, username, platform))
+                return
 
-        if (not "player" in stats):
-            if ("errors" in stats):
-                for error in stats["errors"]:
-                    detail = "Unknown error."
-                    description = "Unknown."
+            if (not "player" in stats):
+                if ("errors" in stats):
+                    for error in stats["errors"]:
+                        detail = "Unknown error."
+                        description = "Unknown."
 
-                    if ("detail" in error):
-                        detail = error["detail"]
+                        if ("detail" in error):
+                            detail = error["detail"]
 
-                    if ("meta" in error and "description" in error["meta"]):
-                        description = error["meta"]["description"]
+                        if ("meta" in error and "description" in error["meta"]):
+                            description = error["meta"]["description"]
 
-                    await self.bot.messaging.reply(ctx.message, "An error occured searching for `{}` on `{}`: {} {}".format(username,
-                        platform,
-                        detail,
-                        description))
+                        await self.bot.messaging.reply(ctx.message, "An error occured searching for `{}` on `{}`: {} {}".format(username,
+                            platform,
+                            detail,
+                            description))
 
-            return
+                return
 
-        stats = stats["player"]
+            stats = stats["player"]
 
         if (stats_selection == "all"):
             stats_option = stats_options[:-1]
@@ -412,8 +419,27 @@ Created at {date}
             success = True
             await self.bot.send_message(ctx.message.channel, embed=embed)
 
+            self.SIEGE_CACHE[username] = {"time": time.time(),
+                                          "stats": stats}
+
         if (not success):
             await self.bot.messaging.reply(ctx.message, "Failed to find `{}` stats for `{}` on `{}`".format(stats_selection, username, platform))
+
+    async def remove_siege_cache(self):
+        while (not self.bot.is_closed):
+            try:
+                siege_cache_copy = self.SIEGE_CACHE.copy()
+                
+                for username, cache in siege_cache_copy.items():
+                    if (time.time() > cache["time"] + self.bot.CONFIG["SIEGE_CACHE_TIME"]):
+                        del self.SIEGE_CACHE[username]
+                
+                siege_cache_copy.clear()
+
+            except Exception as e:
+                self.bot.bot_utils.log_error_to_file(e)
+            
+            await asyncio.sleep(self.bot.CONFIG["SIEGE_CACHE_TIME"] // 2)
 
 def setup(bot):
     bot.add_cog(Utility(bot))
